refactor : NotificationController, NotificationServiceImpl
